backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
cachetools>=5.3.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
PyYAML>=6.0
pyusb>=1.2.1
supabase>=2.0.0
anthropic>=0.75.0
hid>=1.0.5
pydub>=0.25.1
structlog>=24.1.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
requests>=2.31.0
feedparser>=6.0.10
watchdog>=3.0.0
keyboard>=0.13.5
psutil>=5.9.0
# rapidfuzz: Fast fuzzy string matching for game title search
# If unavailable, system gracefully degrades to stdlib difflib
rapidfuzz>=3.0.0
# Optional: enable .7z extraction for PS2 auto-extract
# py7zr>=0.21

# Pinned to avoid ClientOptions.storage regression
supabase==2.23.2
